backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic==2.9.0
pydantic-settings==2.4.0
python-multipart==0.0.9
